+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(slices-xor): xor new function added
 * ✅ test(slices-zip_by): zip_by function's test cases and benchmarks added
 * 🎉 feat(slices-zip_by): zip_by function added
 * 🐛 fix(slices-zip_map_deep): bug fix on go1.17 test files fix